他人の登録情報を削除できないようにした

+++ lib/bookbot/delete.py
@@ -1,6 +1,7 @@
 from lib import get_logger
 from lib.aws.dynamodb import Dynamodb
 from lib.util.converter import Converter
+from lib.util.slack import Slack
 from lib.bookbot.list_history import ListHistory
 
 
@@ -9,6 +10,7 @@ class Delete:
         self.logger = get_logger(__name__)
         self.dynamodb = Dynamodb()
         self.converter = Converter()
+        self.slack = Slack()
         self.list_history = ListHistory()
 
     def specified_entry_no(self, message, entry_no):
@@ -18,16 +20,21 @@ class Delete:
             message.send("対象の登録データが見つかりません")
             return
 
-        text_list = list()
+        # プライマリキー指定なので必ず1件取得
+        item = items[0]
+
+        # 削除できるのは自分の登録情報だけ
+        if item['slack_id'] != self.slack.get_slack_id(message):
+            message.send("削除できるのは自分の登録情報だけです！")
+            return
 
         # レコード削除
         key = dict()
         key['entry_no'] = entry_no
         self.dynamodb.remove(self.dynamodb.default_table, key)
-        text_list.append("以下の登録データを削除しました！")
 
-        # プライマリキー指定なので必ず1件取得
-        item = items[0]
+        text_list = list()
+        text_list.append("以下の登録データを削除しました！")
         text_list.append(self.converter.get_list_str(item))
 
         message.send("\n".join(text_list))
